feat(kandel): Separate geometric Kandel from core Kandel
