+++ test/mocha/config/coverage-tests.json
@@ -1,7 +1,7 @@
 {
   "require": ["ts-node/register", "src/util/test/mochaHooks.ts"],
   "recursive": true,
-  "spec": ["test/unit/*unit.test.ts", "test/integration/*integration.test.ts"],
+  "spec": ["test/unit/**/*.ts", "test/integration/**/*.ts"],
   "timeout": 180000,
   "reporter": "mocha-multi-reporters",
   "reporterOptions": "configFile=test/mocha/config/coverage-tests-multi-reporters.json"
