+++ test/mocha/config/integration-tests.json
@@ -1,7 +1,7 @@
 {
   "require": ["ts-node/register", "src/util/test/mochaHooks.ts"],
   "recursive": true,
-  "spec": "test/integration/*integration.test.ts",
+  "spec": "test/integration/**/*.ts",
   "timeout": 120000,
   "reporter": "mocha-multi-reporters",
   "reporterOptions": "configFile=test/mocha/config/integration-tests-multi-reporters.json"
