+++ test/mocha/config/unit-tests.json
@@ -1,7 +1,7 @@
 {
   "require": ["ts-node/register"],
   "recursive": true,
-  "spec": ["test/unit/*unit.test.ts"],
+  "spec": ["test/unit/**/*.ts"],
   "timeout": 60000,
   "reporter": "mocha-multi-reporters",
   "reporterOptions": "configFile=test/mocha/config/unit-tests-multi-reporters.json"
