+++ test/unit/kandel/kandelDistributionHelper.unit.test.ts
@@ -1,283 +1,13 @@
 import assert from "assert";
 import { Big } from "big.js";
 import { describe, it } from "mocha";
-import KandelDistributionHelper, {
-  TickDistributionParams,
-} from "../../src/kandel/kandelDistributionHelper";
-import { KandelDistribution } from "../../src";
+import KandelDistributionHelper from "../../../src/kandel/kandelDistributionHelper";
+import { KandelDistribution } from "../../../src";
 import {
   assertIsRounded,
   assertSameTicks,
-  createGeneratorStub,
-} from "./kandelDistributionGenerator.unit.test";
-
-describe(`${KandelDistributionHelper.prototype.constructor.name} unit tests suite`, () => {
-  describe(
-    KandelDistributionHelper.prototype.getBaseQuoteTicksFromTick.name,
-    () => {
-      it("can get baseQuoteTicks from tick", () => {
-        // Arrange
-        const sut = new KandelDistributionHelper(4, 6);
-
-        // Act
-        const baseQuoteTicksForAsk = sut.getBaseQuoteTicksFromTick(
-          "asks",
-          2,
-          1000,
-          100,
-          5,
-        );
-        const baseQuoteTicksForBid = sut.getBaseQuoteTicksFromTick(
-          "bids",
-          2,
-          -1000,
-          100,
-          5,
-        );
-
-        // Assert
-        assert.deepStrictEqual(
-          baseQuoteTicksForAsk,
-          [800, 900, 1000, 1100, 1200],
-        );
-        assert.deepStrictEqual(baseQuoteTicksForAsk, baseQuoteTicksForBid);
-      });
-    },
-  );
-
-  describe(
-    KandelDistributionHelper.prototype.calculateBaseQuoteTickOffset.name,
-    () => {
-      it("can calculate based on price ratio", () => {
-        // Arrange
-        const sut = new KandelDistributionHelper(4, 6);
-
-        // Act
-        const actual = sut.calculateBaseQuoteTickOffset(Big(1.08));
-
-        // Assert
-        assert.equal(actual, 769);
-      });
-
-      it("Fails if less than 1", () => {
-        // Arrange
-        const sut = new KandelDistributionHelper(4, 6);
-
-        // Act/assert
-        assert.throws(
-          () => sut.calculateBaseQuoteTickOffset(Big(0.99)),
-          new Error("priceRatio must be larger than 1"),
-        );
-      });
-    },
-  );
-
-  describe(
-    KandelDistributionHelper.prototype.getTickDistributionParams.name,
-    () => {
-      it("calculates sames parameters for all combinations of minPrice, maxPrice, ratio, and pricePoints, and the similar tick-based parameters ", () => {
-        // Arrange
-        const sut = new KandelDistributionHelper(4, 6);
-        const stepSize = 1;
-        const baseQuoteTickOffset = 769;
-        const priceRatio = 1.08;
-        const pricePoints = 7;
-        const maxBaseQuoteTick = 119759;
-        const minBaseQuoteTick = 115145;
-        const midBaseQuoteTick = 117453;
-        const minPrice = Big(1001);
-        const maxPrice = Big(1587.841870262581);
-        const midPrice = Big(1260.971712);
-
-        const expectedParams: TickDistributionParams = {
-          generateFromMid: true,
-          stepSize,
-          baseQuoteTickOffset,
-          maxBaseQuoteTick,
-          minBaseQuoteTick,
-          midBaseQuoteTick,
-          pricePoints,
-        };
-
-        // Act/assert
-        assert.deepStrictEqual(
-          sut.getTickDistributionParams({
-            minPrice,
-            maxPrice,
-            priceRatio,
-            midPrice,
-            stepSize,
-            generateFromMid: true,
-          }),
-          expectedParams,
-        );
-        assert.deepStrictEqual(
-          sut.getTickDistributionParams({
-            minPrice,
-            maxPrice,
-            pricePoints,
-            midPrice,
-            stepSize,
-            generateFromMid: true,
-          }),
-          expectedParams,
-        );
-        assert.deepStrictEqual(
-          sut.getTickDistributionParams({
-            minPrice,
-            priceRatio,
-            pricePoints,
-            midPrice,
-            stepSize,
-            generateFromMid: true,
-          }),
-          expectedParams,
-        );
-        assert.deepStrictEqual(
-          sut.getTickDistributionParams({
-            maxPrice,
-            priceRatio,
-            pricePoints,
-            midPrice,
-            stepSize,
-            generateFromMid: true,
-          }),
-          expectedParams,
-        );
-
-        assert.deepStrictEqual(
-          sut.getTickDistributionParams({
-            minBaseQuoteTick,
-            maxBaseQuoteTick,
-            baseQuoteTickOffset,
-            midBaseQuoteTick,
-            stepSize,
-            generateFromMid: true,
-          }),
-          expectedParams,
-        );
-      });
-
-      it("fails if neither midBaseQuoteTick nor midPrice is given", () => {
-        // Arrange
-        const sut = new KandelDistributionHelper(4, 6);
-
-        // Act/assert
-        assert.throws(
-          () =>
-            sut.getTickDistributionParams({
-              generateFromMid: false,
-              stepSize: 1,
-            }),
-          new Error("midPrice or midBaseQuoteTick must be provided."),
-        );
-      });
-
-      it("can get 2 pricePoints from minPrice and maxPrice", () => {
-        const sut = new KandelDistributionHelper(4, 6);
-
-        // Arrange/Act
-        const params = sut.getTickDistributionParams({
-          minPrice: "1455.3443267746625",
-          maxPrice: "2183.0164901619937",
-          midPrice: "1819.180408468328",
-          pricePoints: 2,
-          stepSize: 1,
-          generateFromMid: false,
-        });
-
-        // Assert
-        assert.equal(
-          params.baseQuoteTickOffset.toString(),
-          params.baseQuoteTickOffset.toFixed(0),
-        );
-      });
-
-      it("throws error if not enough parameters are given", () => {
-        const sut = new KandelDistributionHelper(4, 6);
-        assert.throws(
-          () =>
-            sut.getTickDistributionParams({
-              pricePoints: 10,
-              maxPrice: Big(2),
-              midPrice: Big(1.5),
-              stepSize: 1,
-              generateFromMid: true,
-            }),
-          new Error(
-            "Exactly three of minPrice (or minBaseQuoteTick), maxPrice (or maxBaseQuoteTick), priceRatio (or baseQuoteTickOffset), and pricePoints must be given",
-          ),
-        );
-      });
-
-      it("throws error if only 1 price point", () => {
-        const sut = new KandelDistributionHelper(4, 6);
-        assert.throws(
-          () =>
-            sut.getTickDistributionParams({
-              minPrice: Big(1),
-              maxPrice: Big(2),
-              pricePoints: 1,
-              midPrice: Big(1.5),
-              stepSize: 1,
-              generateFromMid: true,
-            }),
-          new Error("There must be at least 2 price points"),
-        );
-      });
-
-      it("throws if min/max too close", () => {
-        const sut = new KandelDistributionHelper(4, 6);
-        assert.throws(
-          () =>
-            sut.getTickDistributionParams({
-              minPrice: Big(1),
-              maxPrice: Big(1.001),
-              baseQuoteTickOffset: 1000,
-              midPrice: Big(1),
-              stepSize: 1,
-              generateFromMid: true,
-            }),
-          new Error(
-            "minBaseQuoteTick and maxBaseQuoteTick are too close. There must be room for at least two price points",
-          ),
-        );
-      });
-
-      it("throws if min too low", () => {
-        const sut = new KandelDistributionHelper(4, 6);
-        assert.throws(
-          () =>
-            sut.getTickDistributionParams({
-              minBaseQuoteTick: -1e10,
-              maxPrice: Big(1.001),
-              baseQuoteTickOffset: 1000,
-              midPrice: Big(1),
-              stepSize: 1,
-              generateFromMid: true,
-            }),
-          new Error("minBaseQuoteTick too low."),
-        );
-      });
-
-      it("throws if max too high", () => {
-        const sut = new KandelDistributionHelper(4, 6);
-        assert.throws(
-          () =>
-            sut.getTickDistributionParams({
-              minBaseQuoteTick: 1,
-              maxBaseQuoteTick: 1e10,
-              baseQuoteTickOffset: 1000,
-              midPrice: Big(1),
-              stepSize: 1,
-              generateFromMid: true,
-            }),
-          new Error("maxBaseQuoteTick too high."),
-        );
-      });
-    },
-  );
-});
+} from "./generalKandelDistributionGenerator.unit.test";
+import { createGeneratorStub } from "./geometricKandel/geometricKandelDistributionGenerator.unit.test";
 
 describe(`${KandelDistributionHelper.prototype.constructor.name} unit tests suite`, () => {
   describe(
@@ -489,7 +219,7 @@ describe(`${KandelDistributionHelper.prototype.constructor.name} unit tests suit
       let sut: KandelDistributionHelper;
       beforeEach(async () => {
         const generator = createGeneratorStub();
-        sut = generator.distributionHelper;
+        sut = generator.generalDistributionHelper;
         distribution = await generator.calculateDistribution({
           distributionParams: {
             minPrice: Big(1000),
@@ -519,10 +249,11 @@ describe(`${KandelDistributionHelper.prototype.constructor.name} unit tests suit
         });
 
         // Assert
-        assertSameTicks(result.distribution, distribution);
+        assertSameTicks(result.distribution.wrappedDistribution, distribution);
 
         const oldVolume = distribution.getOfferedVolumeForDistribution();
-        const newVolume = result.distribution.getOfferedVolumeForDistribution();
+        const newVolume =
+          result.distribution.wrappedDistribution.getOfferedVolumeForDistribution();
         assert.equal(
           newVolume.requiredBase.toNumber(),
           oldVolume.requiredBase.add(baseDelta).toNumber(),
@@ -532,31 +263,35 @@ describe(`${KandelDistributionHelper.prototype.constructor.name} unit tests suit
           oldVolume.requiredQuote.add(quoteDelta).toNumber(),
         );
 
-        assertIsRounded(result.distribution);
+        assertIsRounded(result.distribution.wrappedDistribution);
         assert.equal(result.totalBaseChange.toNumber(), baseDelta.toNumber());
         assert.equal(result.totalQuoteChange.toNumber(), quoteDelta.toNumber());
 
-        result.distribution.getLiveOffers("asks").forEach((o) => {
-          assert.ok(o.gives.gte(Big(1)), "ask base should be above minimum");
-          assert.ok(
-            Big(
-              sut.askTickPriceHelper.inboundFromOutbound(o.tick, o.gives),
-            ).gte(Big(9000)),
-            "ask quote should be above minimum",
-          );
-        });
-        result.distribution.getLiveOffers("bids").forEach((o) => {
-          assert.ok(
-            o.gives.gte(Big(9000)),
-            "bid quote should be above minimum",
-          );
-          assert.ok(
-            sut.bidTickPriceHelper
-              .inboundFromOutbound(o.tick, o.gives)
-              .gte(Big(1)),
-            "bid base should be above minimum",
-          );
-        });
+        result.distribution.wrappedDistribution
+          .getLiveOffers("asks")
+          .forEach((o) => {
+            assert.ok(o.gives.gte(Big(1)), "ask base should be above minimum");
+            assert.ok(
+              Big(
+                sut.askTickPriceHelper.inboundFromOutbound(o.tick, o.gives),
+              ).gte(Big(9000)),
+              "ask quote should be above minimum",
+            );
+          });
+        result.distribution.wrappedDistribution
+          .getLiveOffers("bids")
+          .forEach((o) => {
+            assert.ok(
+              o.gives.gte(Big(9000)),
+              "bid quote should be above minimum",
+            );
+            assert.ok(
+              sut.bidTickPriceHelper
+                .inboundFromOutbound(o.tick, o.gives)
+                .gte(Big(1)),
+              "bid base should be above minimum",
+            );
+          });
       });
 
       [
@@ -580,7 +315,7 @@ describe(`${KandelDistributionHelper.prototype.constructor.name} unit tests suit
 
           // Assert
           const newVolume =
-            result.distribution.getOfferedVolumeForDistribution();
+            result.distribution.wrappedDistribution.getOfferedVolumeForDistribution();
           assert.equal(
             newVolume.requiredBase.toNumber(),
             oldVolume.requiredBase.add(baseDelta ?? Big(0)).toNumber(),
